move filter-value related code out of __init__.py
